requirements.txt
asgiref==3.8.1
autopep8==2.3.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-admin-adminlte==1.0.8
django-adminlte-ui==2.1.1
django-adminlte3==0.1.6
django-channels==0.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.17.0
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gprof2dot==2024.6.6
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httpie==3.2.4
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
qrcode==8.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
